+++ tests/LinkDotNet.StringBuilder.UnitTests/ValueStringBuilderTests.cs
@@ -416,4 +416,40 @@ public class ValueStringBuilderTests
 
         sb.ToString().Should().Be("Hello World");
     }
+
+    [Fact]
+    public void ConcatArbitraryValues()
+    {
+        var result = ValueStringBuilder.Concat(new[] { "Hello", " ", "World" });
+
+        result.Should().Be("Hello World");
+    }
+
+    [Fact]
+    public void ShouldReturnEmptyStringIfEmptyArray()
+    {
+        var result = ValueStringBuilder.Concat(Array.Empty<string>());
+
+        result.Should().Be(string.Empty);
+    }
+
+    [Fact]
+    public void ConcatBooleanWithNumber()
+    {
+        var result = ValueStringBuilder.Concat(true, 1);
+
+        result.Should().Be("True1");
+    }
+
+    [Theory]
+    [InlineData("Hello", true)]
+    [InlineData("Hallo", false)]
+    public void GivenReadOnlySpan_WhenCallingEquals_ThenReturningWhenEqual(string input, bool expected)
+    {
+        using var builder = new ValueStringBuilder("Hello");
+
+        var isEqual = builder.Equals(input);
+
+        isEqual.Should().Be(expected);
+    }
 }